requirements.txt
discord.py 
python-dotenv 
Pillow 
requests
pyautogui
